Bug fix in update log.

+++ puppet_master.py
@@ -63,7 +63,8 @@ class PuppetMaster(object):
         self._updating = True
         self.stop()
 
-        os.remove(self.config.update.logfile)
+        if os.path.exists(self.config.update.logfile):
+            os.remove(self.config.update.logfile)
         success = check_call(['poppy-update'])
 
         self.start()
